+++ frontend/src/pages/Projects/Project/ChangeRequest/tableConfig.tsx
@@ -1,10 +1,15 @@
 import { GridColDef } from "@mui/x-data-grid";
-import { dateFormatter as valueFormatter } from "utils/formatDate";
+import { dateFormatter } from "utils/formatDate";
 
 export const tableConfig = () => {
   const defaultFlex = 3;
   const tableColumns: GridColDef[] = [
-    { field: "init_date", headerName: "Initiation Date", valueFormatter, flex: defaultFlex },
+    {
+      field: "init_date",
+      headerName: "Initiation Date",
+      valueFormatter: dateFormatter,
+      flex: defaultFlex,
+    },
     { field: "summary", headerName: "Summary", flex: defaultFlex },
     { field: "types", headerName: "Types", flex: defaultFlex },
     { field: "version", headerName: "Version", flex: defaultFlex },