+++ frontend/src/pages/Projects/Project/ProjectDetails/Deliverables/tableConfig.ts
@@ -1,5 +1,5 @@
 import { GridColDef } from "@mui/x-data-grid";
-import { dateFormatter as valueFormatter } from "utils/formatDate";
+import { dateFormatter } from "utils/formatDate";
 
 export const tableConfig = () => {
   const defaultFlex = 3;
@@ -10,8 +10,18 @@ export const tableConfig = () => {
       flex: defaultFlex,
     },
     { field: "description", headerName: "Description", flex: defaultFlex },
-    { field: "start_date", headerName: "Start Date", valueFormatter, flex: defaultFlex },
-    { field: "completion_date", headerName: "Completion Date", valueFormatter, flex: defaultFlex },
+    {
+      field: "start_date",
+      headerName: "Start Date",
+      valueFormatter: dateFormatter,
+      flex: defaultFlex,
+    },
+    {
+      field: "completion_date",
+      headerName: "Completion Date",
+      valueFormatter: dateFormatter,
+      flex: defaultFlex,
+    },
     { field: "deliverable_amount", headerName: "Deliverable Amount", flex: defaultFlex },
     { field: "recoverable_amount", headerName: "Recoverable Amount", flex: defaultFlex },
     { field: "project_number", headerName: "Project Number", flex: defaultFlex },